code/preprocess/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
atomicwrites=1.4.0=pypi_0
attrs=20.2.0=py_0
beautifulsoup4=4.9.3=pypi_0
blas=2.17=openblas
blis=0.7.4=pypi_0
boto3=1.17.53=pypi_0
botocore=1.20.53=pypi_0
bounded-pool-executor=0.0.3=pypi_0
c-compiler=1.1.1=haf1e3a3_0
ca-certificates=2020.7.22=0
catalogue=1.0.0=pypi_0
cctools=949.0.1=h22b1bf0_1
cctools_osx-64=949.0.1=h5ba7a2e_1
certifi=2020.6.20=py38_0
chardet=4.0.0=pypi_0
clang=10.0.0=default_hf57f61e_0
clang_osx-64=10.0.0=h05bbb7f_0
clangxx=10.0.0=default_hf57f61e_0
clangxx_osx-64=10.0.0=h05bbb7f_1
click=7.1.2=pypi_0
compiler-rt=10.0.0=h47ead80_0
compiler-rt_osx-64=10.0.0=hbcc88fd_0
compilers=1.1.1=0
coverage=4.5.1=pypi_0
coveralls=1.3.0=pypi_0
cxx-compiler=1.1.1=h879752b_0
cycler=0.10.0=pypi_0
cymem=2.0.5=pypi_0
cython=0.29.21=py38hb1e8313_0
datefinder=0.7.1=pypi_0
dateparser=1.0.0=pypi_0
decorator=4.4.2=pypi_0
docopt=0.6.2=pypi_0
en-core-web-lg=2.3.1=pypi_0
en-core-web-md=2.3.1=pypi_0
en-core-web-sm=3.0.0=pypi_0
fortran-compiler=1.1.1=he991be0_0
geometric=0.9.7.2=pypi_0
gfortran_impl_osx-64=7.3.0=hf4212f2_2
gfortran_osx-64=7.3.0=h22b1bf0_11
gmp=6.1.2=hb37e062_1
idna=2.10=pypi_0
iniconfig=1.0.1=py_0
isl=0.19=0
jinja2=2.11.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=0.16.0=py_0
kiwisolver=1.3.1=pypi_0
ld64=530=1
ld64_osx-64=530=h3c32e8a_1
libblas=3.8.0=17_openblas
libcblas=3.8.0=17_openblas
libcxx=10.0.1=h5f48129_0
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=4.0.0=h2d743fc_11
libgfortran4=7.5.0=h2d743fc_11
libiconv=1.16=h1de35cc_0
liblapack=3.8.0=17_openblas
liblapacke=3.8.0=17_openblas
libllvm10=10.0.0=h21ff451_2
libopenblas=0.3.10=openmp_h63d9170_4
llvm-openmp=10.0.1=h28b9765_0
llvm-tools=10.0.0=h21ff451_2
lxml=4.6.3=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.4.2=pypi_0
more-itertools=8.5.0=py_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
murmurhash=1.0.5=pypi_0
mwparserfromhell=0.6=pypi_0
ncurses=6.2=h0a44026_1
networkx=2.5.1=pypi_0
neuralcoref=4.0=dev_0
nltk=3.5=pypi_0
numpy=1.20.2=pypi_0
openssl=1.1.1h=haf1e3a3_0
packaging=20.4=py_0
pathy=0.4.0=pypi_0
pillow=8.2.0=pypi_0
pip=20.2.3=py38_0
plac=1.1.3=pypi_0
pluggy=0.13.1=py38_0
pqdm=0.1.0=pypi_0
preshed=3.0.5=pypi_0
protobuf=3.16.0=pypi_0
py=1.9.0=py_0
pydantic=1.7.3=pypi_0
pyparsing=2.4.7=py_0
pytest=3.7.1=pypi_0
pytest-cov=2.5.1=pypi_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=pypi_0
python_abi=3.8=1_cp38
pytz=2021.1=pypi_0
readline=8.0=h1de35cc_0
regex=2021.4.4=pypi_0
requests=2.25.1=pypi_0
requests-toolbelt=0.9.1=pypi_0
s3transfer=0.3.7=pypi_0
scipy=1.5.2=py38h1402333_0
setuptools=49.6.0=py38_1
six=1.15.0=py_0
smart-open=3.0.0=pypi_0
soupsieve=2.2.1=pypi_0
spacy=2.3.5=pypi_0
spacy-legacy=3.0.2=pypi_0
sqlite=3.33.0=hffcf06c_0
srsly=1.0.5=pypi_0
stanza=1.2=pypi_0
subword-nmt=0.3.7=pypi_0
tapi=1000.10.8=ha1b3eb9_0
thinc=7.4.5=pypi_0
timefhuman=0.0.5=pypi_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.1=py_0
torch=1.8.1=pypi_0
tqdm=4.60.0=pypi_0
typer=0.3.2=pypi_0
typing-extensions=3.10.0.0=pypi_0
tzlocal=2.1=pypi_0
urllib3=1.26.4=pypi_0
utils=1.0.1=pypi_0
wasabi=0.8.2=pypi_0
wheel=0.35.1=py_0
wikimapper=0.1.5=pypi_0
wikipedia=1.4.0=pypi_0
xmltodict=0.12.0=pypi_0
xmltojson=0.2.0=pypi_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
